backend/requirements.txt

fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart